Explicitly set initsystem to `systemd`
libreswan autodetects a `systemd` initsystem by verifying that
`/proc/1/comm` has the value "systemd". According to the release notes
for systemd v248:

  The manager may be configured at compile time to use the fexecve()
  instead of the execve() system call when spawning processes. Using
  fexecve() closes a window between checking the security context of an
  executable and spawning it, but unfortunately the kernel displays
  stale information in the process' "comm" field, which impacts ps
  output and such.

The Clear Linux OS systemd package (currently at v249.1) appears to not
use fexecve() -- it is listed as a "disabled feature" -- but
/proc/1/comm is observed to have a different value when running a build
with systemd-nspawn 249: `(sd-stubinit)`. When running the command `make
build MOCK_OPTS=--old-chroot` on a system running systemd 249, *or*
running `make build` -- assuming the default configuration of mock using
systemd-nspawn for builds -- on a system with systemd 246 installed
(release < 34900), the value is `systemd`.